% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mws_utilities.R
\name{mwsPage}
\alias{mwsPage}
\title{Creates a selected multiwindow shiny page.}
\usage{
mwsPage(ui)
}
\arguments{
\item{ui}{a selection from ui_list, the list of shiny ui pages.}
}
\value{
user interface for the selected multi-window page
}
\description{
Creates a selected multiwindow shiny page.
}
