% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mws_utilities.R
\name{mwsSelectorPage}
\alias{mwsSelectorPage}
\title{Creates the "Window Selector" page.}
\usage{
mwsSelectorPage(win_titles)
}
\arguments{
\item{win_titles}{vector of uniquely named strings, corresponding to window titles. Must be same length as ui_win, and titles must be same index as corresponding ui page in ui_win. No windows can be named 'WindowSelector', and titles cannot have spaces.}
}
\value{
user interface for window selector page
}
\description{
Creates the "Window Selector" page.
}
