% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mws_utilities.R
\name{serverFunct}
\alias{serverFunct}
\title{Render server output.}
\usage{
serverFunct(serverValues, output, serv_out_list)
}
\arguments{
\item{serverValues}{traditional Shiny inputs, as well as user-created values (reactive values), to be used in rendering output}

\item{output}{traditional shiny output}

\item{serv_out_list}{a named list of functions that render output. Each function is of the form function(calc), where calc is a named list containing the traditional shiny input and reactive values that have calculated values derived from input. It returns the results of a shiny render function. The name of each function corresponds to its output label.}
}
\value{
traditional shiny output argument
}
\description{
Render server output.
}
