% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mws_utilities.R
\name{mwsUI}
\alias{mwsUI}
\title{Renders user interface for all mwshiny windows.}
\usage{
mwsUI(ui_list)
}
\arguments{
\item{ui_list}{named list of shiny UI pages. The name of each entry in the UI page list corresponds to its window title. No windows can be named 'WindowSelector', titles must be uniquely named, and titles cannot have spaces.}
}
\value{
ui: user interfaces for all windows
}
\description{
Renders user interface for all mwshiny windows.
}
